pyproject.toml
[project]
name = "appcraft"
version = "0.5.3"
description = "Appcraft is a modular framework designed to simplify software development by providing a structured and extensible architecture. It enables developers to create and manage complex applications with ease, ensuring scalability and maintainability."
authors = [{ name = "Dux Tecnologia", email = "contato@tpereira.com.br" }]
dependencies = [
    "prompt_toolkit",
    "toml",
    "py-automapper (>=2.1.0,<3.0.0)"
]
classifiers= [
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent'
]

requires-python = ">=3.8"

readme = {file = "README.md", content-type = "text/markdown"}

[project.scripts]
appcraft = "appcraft.scripts.cli:main"


[build-system]
requires = ["setuptools", "wheel", "toml"]
build-backend = "setuptools.build_meta"
